refactor: move all string styling work to `iroh-ctl`
