+++ iroh-rpc-client/src/lib.rs
@@ -6,4 +6,4 @@ mod store;
 
 pub use crate::client::Client;
 pub use crate::client::RpcClientConfig;
-pub use crate::status::ServiceStatus;
+pub use crate::status::{ServiceStatus, StatusRow, StatusTable};
