+++ iroh-rpc-types/src/lib.rs
@@ -2,5 +2,6 @@
 pub mod gateway;
 pub mod p2p;
 pub mod store;
+#[cfg(feature = "testing")]
 #[allow(clippy::all)]
 pub mod test;
